collate tests
